requirements.txt
streamlit
yfinance
google-generativeai
finnhub-python
pandas